juridica_flow/requirements.txt
fastapi>=0.112
uvicorn>=0.30
SQLAlchemy>=2.0
pydantic>=2.6
python-dotenv>=1.0
psycopg2-binary>=2.9
Jinja2>=3.1
python-multipart>=0.0.9
